+++ src/linearmixedmodel.jl
@@ -177,7 +177,6 @@ function LinearMixedModel(
     θ = foldl(vcat, getθ(c) for c in reterms)
     optsum = OptSummary(θ, lbd)
     optsum.sigma = isnothing(σ) ? nothing : T(σ)
-    fill!(optsum.xtol_abs, 1.0e-10)
     return LinearMixedModel(
         form,
         reterms,
@@ -206,21 +205,16 @@ function StatsAPI.fit(
     )
 end
 
-function StatsAPI.fit(
-    ::Type{LinearMixedModel},
+function StatsAPI.fit(::Type{LinearMixedModel},
     f::FormulaTerm,
     tbl::Tables.ColumnTable;
     wts=[],
     contrasts=Dict{Symbol,Any}(),
-    progress=true,
-    REML=false,
     σ=nothing,
-    thin=typemax(Int),
     amalgamate=true,
-)
-    return fit!(
-        LinearMixedModel(f, tbl; contrasts, wts, σ, amalgamate); progress, REML, thin
-    )
+    kwargs...)
+    lmm = LinearMixedModel(f, tbl; contrasts, wts, σ, amalgamate)
+    return fit!(lmm; kwargs...)
 end
 
 function _offseterr()
@@ -433,14 +427,15 @@ end
 """
     fit!(m::LinearMixedModel; progress::Bool=true, REML::Bool=m.optsum.REML,
                               σ::Union{Real, Nothing}=m.optsum.sigma,
-                              thin::Int=typemax(Int))
+                              thin::Int=typemax(Int),
+                              fitlog::Bool=true)
 
 Optimize the objective of a `LinearMixedModel`.  When `progress` is `true` a
 `ProgressMeter.ProgressUnknown` display is shown during the optimization of the
 objective, if the optimization takes more than one second or so.
 
-At every `thin`th iteration  is recorded in `fitlog`, optimization progress is
-saved in `m.optsum.fitlog`.
+The `thin` argument is ignored: it had no impact on the final model fit and the logic around
+thinning the `fitlog` was needlessly complicated for a trivial performance gain.
 """
 function StatsAPI.fit!(
     m::LinearMixedModel{T};
@@ -448,6 +443,9 @@ function StatsAPI.fit!(
     REML::Bool=m.optsum.REML,
     σ::Union{Real,Nothing}=m.optsum.sigma,
     thin::Int=typemax(Int),
+    fitlog::Bool=false,
+    backend::Symbol=m.optsum.backend,
+    optimizer::Symbol=m.optsum.optimizer,
 ) where {T}
     optsum = m.optsum
     # this doesn't matter for LMM, but it does for GLMM, so let's be consistent
@@ -459,55 +457,29 @@ function StatsAPI.fit!(
             ArgumentError("The response is constant and thus model fitting has failed")
         )
     end
-    opt = Opt(optsum)
     optsum.REML = REML
     optsum.sigma = σ
-    prog = ProgressUnknown(; desc="Minimizing", showspeed=true)
-    # start from zero for the initial call to obj before optimization
-    iter = 0
-    fitlog = optsum.fitlog
-    function obj(x, g)
-        isempty(g) || throw(ArgumentError("g should be empty for this objective"))
-        iter += 1
-        val = if isone(iter) && x == optsum.initial
-            optsum.finitial
-        else
-            try
-                objective(updateL!(setθ!(m, x)))
-            catch ex
-                # This can happen when the optimizer drifts into an area where
-                # there isn't enough shrinkage. Why finitial? Generally, it will
-                # be the (near) worst case scenario value, so the optimizer won't
-                # view it as an optimum. Using Inf messes up the quadratic
-                # approximation in BOBYQA.
-                ex isa PosDefException || rethrow()
-                optsum.finitial
-            end
-        end
-        progress && ProgressMeter.next!(prog; showvalues=[(:objective, val)])
-        !isone(iter) && iszero(rem(iter, thin)) && push!(fitlog, (copy(x), val))
-        return val
-    end
-    NLopt.min_objective!(opt, obj)
+    optsum.backend = backend
+    optsum.optimizer = optimizer
+
     try
         # use explicit evaluation w/o calling opt to avoid confusing iteration count
-        optsum.finitial = objective(updateL!(setθ!(m, optsum.initial)))
+        optsum.finitial = objective!(m, optsum.initial)
     catch ex
         ex isa PosDefException || rethrow()
         # give it one more try with a massive change in scaling
         @info "Initial objective evaluation failed, rescaling initial guess and trying again."
         @warn """Failure of the initial evaluation is often indicative of a model specification
-                 that is not well supported by the data and/or a poorly scaled model.
-              """
+                that is not well supported by the data and/or a poorly scaled model.
+            """
         optsum.initial ./=
             (isempty(m.sqrtwts) ? 1.0 : maximum(m.sqrtwts)^2) *
             maximum(response(m))
-        optsum.finitial = objective(updateL!(setθ!(m, optsum.initial)))
+        optsum.finitial = objective!(m, optsum.initial)
     end
-    empty!(fitlog)
-    push!(fitlog, (copy(optsum.initial), optsum.finitial))
-    fmin, xmin, ret = NLopt.optimize!(opt, copyto!(optsum.final, optsum.initial))
-    ProgressMeter.finish!(prog)
+
+    xmin, fmin = optimize!(m; progress, fitlog)
+
     ## check if small non-negative parameter values can be set to zero
     xmin_ = copy(xmin)
     lb = optsum.lowerbd
@@ -516,24 +488,22 @@ function StatsAPI.fit!(
             xmin_[i] = zero(T)
         end
     end
-    loglength = length(fitlog)
     if xmin ≠ xmin_
-        if (zeroobj = obj(xmin_, T[])) ≤ (fmin + optsum.ftol_zero_abs)
+        if (zeroobj = objective!(m, xmin_)) ≤ (fmin + optsum.ftol_zero_abs)
             fmin = zeroobj
             copyto!(xmin, xmin_)
-        elseif length(fitlog) > loglength
-            # remove unused extra log entry
-            pop!(fitlog)
+            fitlog && push!(optsum.fitlog, (copy(xmin), fmin))
         end
     end
-    ## ensure that the parameter values saved in m are xmin
-    updateL!(setθ!(m, xmin))
 
-    optsum.feval = opt.numevals
+    # unlike GLMM we don't need to populate optsum.finitial here
+    # because we do that during the initial guess and rescale check
+
+    ## ensure that the parameter values saved in m are xmin
+    objective!(m)(xmin)
+
     optsum.final = xmin
     optsum.fmin = fmin
-    optsum.returnvalue = ret
-    _check_nlopt_return(ret)
     return m
 end
 
@@ -826,25 +796,6 @@ function objective(m::LinearMixedModel{T}) where {T}
         muladd(denomdf, muladd(2, log(σ), log2π), (logdet(m) + pwrss(m) / σ^2))
     end
     return isempty(wts) ? val : val - T(2.0) * sum(log, wts)
-end
-
-"""
-    objective!(m::LinearMixedModel, θ)
-    objective!(m::LinearMixedModel)
-
-Equivalent to `objective(updateL!(setθ!(m, θ)))`.
-
-When `m` has a single, scalar random-effects term, `θ` can be a scalar.
-
-The one-argument method curries and returns a single-argument function of `θ`.
-
-Note that these methods modify `m`.
-The calling function is responsible for restoring the optimal `θ`.
-"""
-function objective! end
-
-function objective!(m::LinearMixedModel)
-    return Base.Fix1(objective!, m)
 end
 
 function objective!(m::LinearMixedModel{T}, θ) where {T}